Freeo Features Query Service Initial Commit
This is a simple standalone golang version of a query service with a dummy implementation of the functions that fetch
data from a data source and it returns static data. This version will need some changes to run as an AWS Lambda, but
as a standalone for proof of concept it can be compiled and run. There are also some unit tests (basic as they may be).